requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
requests==2.32.3
pytz